made code level graph name changes
