+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?><!-- Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license at the bottom of this file. -->
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.microsoft.office365.connectunified">
+    package="com.microsoft.office365.connectmicrosoftgraph">
 
     <!-- Required to connect to Office 365 -->
     <uses-permission android:name="android.permission.INTERNET" />
@@ -11,7 +11,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name="com.microsoft.office365.connectunified.ConnectActivity"
+            android:name="com.microsoft.office365.connectmicrosoftgraph.ConnectActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait">
             <intent-filter>
@@ -21,13 +21,13 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.microsoft.office365.connectunified.SendMailActivity"
+            android:name="com.microsoft.office365.connectmicrosoftgraph.SendMailActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
-            android:parentActivityName="com.microsoft.office365.connectunified.ConnectActivity">
+            android:parentActivityName="com.microsoft.office365.connectmicrosoftgraph.ConnectActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.microsoft.office365.connectunified.ConnectActivity" />
+                android:value="com.microsoft.office365.connectmicrosoftgraph.ConnectActivity" />
         </activity>
         <activity
             android:name="com.microsoft.aad.adal.AuthenticationActivity"
