+++ app/src/main/java/com/microsoft/office365/connectmicrosoftgraph/Constants.java
@@ -1,16 +1,17 @@
 /*
  *  Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license at the bottom of this file.
  */
-package com.microsoft.office365.connectunified.vo;
+package com.microsoft.office365.connectmicrosoftgraph;
 
-import com.google.gson.annotations.SerializedName;
-
-public class ToRecipientsVO {
-
-    @SerializedName("EmailAddress")
-    public EmailAddressVO emailAddress;
-
+interface Constants {
+    String AUTHORITY_URL = "https://login.microsoftonline.com/common";
+    // Update these two constants with the values for your application:
+    String CLIENT_ID = "ENTER_YOUR_CLIENT_ID";
+    String REDIRECT_URI = "ENTER_YOUR_REDIRECT_URI";
+    String MICROSOFT_GRAPH_API_ENDPOINT = "https://graph.microsoft.com/v1.0/";
+    String MICROSOFT_GRAPH_API_ENDPOINT_RESOURCE_ID = "https://graph.microsoft.com/";
 }
+
 // *********************************************************
 //
 // O365-Android-Microsoft-Graph-Connect, https://github.com/OfficeDev/O365-Android-Microsoft-Graph-Connect
