+++ app/src/main/java/com/microsoft/office365/connectmicrosoftgraph/vo/BodyVO.java
@@ -1,14 +1,17 @@
 /*
  *  Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license at the bottom of this file.
  */
-package com.microsoft.office365.connectunified.vo;
+package com.microsoft.office365.connectmicrosoftgraph.vo;
 
 import com.google.gson.annotations.SerializedName;
 
-public class EmailAddressVO {
+public class BodyVO {
 
-    @SerializedName("Address")
-    public String mAddress;
+    @SerializedName("ContentType")
+    public String mContentType;
+
+    @SerializedName("Content")
+    public String mContent;
 
 }
 // *********************************************************
