+++ app/src/main/res/menu/send_mail.xml
@@ -1,7 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
 <!-- Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license at the bottom of this file. -->
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.microsoft.office365.connectunified.SendMail`Activity">
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
         android:id="@+id/disconnectMenuitem"
         android:title="@string/disconnect_menu_item_text"></item>
